[TICKET-90]Account, Jwt로 로그인 방식 변경

+++ src/main/java/com/ticket/captain/festivalDetail/FestivalDetail.java
@@ -8,6 +8,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity(name = "festival_detail")
@@ -30,7 +31,7 @@ public class FestivalDetail extends BaseEntity {
     private Long amount;
 
     @Column(name = "ticket_price")
-    private Long price;
+    private BigDecimal price;
 
     @Column(name = "perform_date")
     private LocalDateTime processDate;
@@ -39,7 +40,7 @@ public class FestivalDetail extends BaseEntity {
     private LocalDateTime drawDate;
 
     @Builder
-    public FestivalDetail(String salesType,Long amount, Long price,
+    public FestivalDetail(String salesType,Long amount, BigDecimal price,
                           LocalDateTime processDate, LocalDateTime drawDate, Festival festival) {
         this.salesType = salesType;
         this.amount = amount;
@@ -50,7 +51,7 @@ public class FestivalDetail extends BaseEntity {
     }
 
 
-    public void update(String salesType,Long amount, Long price,
+    public void update(String salesType,Long amount, BigDecimal price,
                        LocalDateTime processDate, LocalDateTime drawDate) {
         this.amount = amount;
         this.price = price;
